requirements.txt
yt-dlp
google-genai
pandas
pydantic
datasets
python-dotenv
pyarrow
